Refactor to buld common error links data from its object
